backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-socketio==5.3.6
flask-jwt-extended==4.6.0
flask-mail==0.10.0
pymongo==4.7.1
python-dotenv==1.0.1
google-generativeai==0.8.6
eventlet==0.36.1
bcrypt==4.1.3
python-dateutil==2.9.0
geopy==2.4.1
requests==2.32.3
gunicorn==22.0.0
google-genai==1.7.0
